@@ -35,19 +35,19 @@
   },
   "main": "public/electron.js",
   "dependencies": {
-    "@emotion/react": "^11.4.1",
-    "@emotion/styled": "^11.3.0",
-    "@mui/icons-material": "^5.0.1",
-    "@mui/material": "^5.0.2",
-    "crc": "^3.8.0",
+    "@emotion/react": "^11.7.1",
+    "@emotion/styled": "^11.6.0",
+    "@mui/icons-material": "^5.2.5",
+    "@mui/material": "^5.2.7",
+    "crc": "^4.1.0",
     "cross-env": "^7.0.3",
     "electron-is-dev": "^2.0.0",
     "react": "^17.0.2",
     "react-contexify": "^5.0.0",
     "react-dom": "^17.0.2",
-    "react-router": "^5.2.1",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "^4.0.3"
+    "react-router": "^6.2.1",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "^5.0.0"
   },
   "scripts": {
     "react-start": "react-scripts start",
@@ -72,15 +72,16 @@
     ]
   },
   "devDependencies": {
-    "concurrently": "^6.2.2",
-    "electron": "^15.0.0",
-    "electron-builder": "^22.11.7",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-plugin-import": "^2.24.2",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.26.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "concurrently": "^7.0.0",
+    "electron": "^16.0.6",
+    "electron-builder": "^22.14.5",
+    "eslint": "^8.6.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
     "wait-on": "^6.0.0"
-  }
+  },
+  "packageManager": "yarn@3.1.1"
 }
