@@ -1,17 +1,17 @@
 import React from 'react';
 import Button from '@mui/material/Button';
 import ArrowLeftIcon from '@mui/icons-material/KeyboardArrowLeft';
-import { useHistory } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 
 const BackButton = () => {
-  const history = useHistory();
+  const navigate = useNavigate();
   /**
    * Call the goBack event
    * @param e The event argument
    */
   const back = (e) => {
     if (e) e.preventDefault();
-    history.goBack();
+    navigate(-1);
   };
 
   return (
